# Optimized byDchar