```
$ ldc2 -O -release test.d && ./test
old		1 sec, 664 ms, 214 μs, and 8 hnsecs
new		1 sec, 176 ms, 790 μs, and 3 hnsecs

$ dmd -O -release -inline test.d && ./test
old		3 secs, 281 ms, 757 μs, and 1 hnsec
new		2 secs, 757 ms, 328 μs, and 1 hnsec
```

1. Removes extra call to encode
2. Removes bounds checks on the static array and two `ushort` member variables
3. In order to change the layout, this change uses a different type all together, thereby avoiding tons of  hard to follow `static if`s inside the original type

code:

<details>

```d
import std.stdio;
import std.exception;
import std.conv;
import std.ascii;
import std.range;
import std.traits;
import std.typecons;
import std.string;
import std.datetime.systime;
import std.datetime.date;
import std.datetime.timezone;
import std.datetime.stopwatch;
import core.time;
import std.random;
import std.utf;

enum testCount = 10_000_000;
__gshared string a = "test😀😁😂😃😀😁😂😃😀😁😂😃";
__gshared size_t res;

template byUTF1(C)
if (isSomeChar!C)
{
    static if (!is(Unqual!C == C))
        alias byUTF1 = byUTF1!(Unqual!C);
    else:

    auto ref byUTF1(R)(R r)
        if (isAutodecodableString!R && isInputRange!R && isSomeChar!(ElementEncodingType!R))
    {
        return byUTF1(r.byCodeUnit());
    }

    auto ref byUTF1(R)(R r)
        if (!isAutodecodableString!R && isInputRange!R && isSomeChar!(ElementEncodingType!R))
    {
        alias RC = Unqual!(ElementEncodingType!R);

        static if (is(RC == C))
        {
            return r.byCodeUnit();
        }
        else
        {
            static struct Result
            {
                @property bool empty()
                {
                    return pos == fill && r.empty;
                }

                @property auto front() scope // 'scope' required by call to decodeFront() below
                {
                    if (pos == fill)
                    {
                        pos = 0;
                        auto c = r.front;

                        if (c <= 0x7F)
                        {
                            fill = 1;
                            r.popFront;
                            buf[pos] = cast(C) c;
                        }
                        else
                        {
                            static if (is(RC == dchar))
                            {
                                r.popFront;
                                dchar dc = c;
                            }
                            else
                                dchar dc = () @trusted { return decodeFront!(Yes.useReplacementDchar)(r); }();
                            fill = cast(ushort) encode!(Yes.useReplacementDchar)(buf, dc);
                        }
                    }
                    return buf[pos];
                }

                void popFront()
                {
                    if (pos == fill)
                        front;
                    ++pos;
                }

                static if (isForwardRange!R)
                {
                    @property auto save() return scope
                    /* `return scope` cannot be inferred because compiler does not
                     * track it backwards from assignment to local `ret`
                     */
                    {
                        auto ret = this;
                        ret.r = r.save;
                        return ret;
                    }
                }

            private:

                R r;
                C[4 / C.sizeof] buf = void;
                ushort pos, fill;
            }

            return Result(r);
        }
    }
}

template byUTF2(C)
if (isSomeChar!C)
{
    static if (!is(Unqual!C == C))
        alias byUTF2 = byUTF2!(Unqual!C);
    else:

    auto ref byUTF2(R)(R r)
        if (isAutodecodableString!R && isInputRange!R && isSomeChar!(ElementEncodingType!R))
    {
        return byUTF2(r.byCodeUnit());
    }

    auto ref byUTF2(R)(R r)
        if (!isAutodecodableString!R && isInputRange!R && isSomeChar!(ElementEncodingType!R))
    {
        alias RC = Unqual!(ElementEncodingType!R);

        static if (is(RC == C))
        {
            return r.byCodeUnit();
        }
        else static if (is(C == dchar))
        {
            static struct Result
            {
                this(R val)
                {
                    r = val;
                    popFront();
                }

                @property bool empty()
                {
                    return buff == uint.max;
                }

                @property auto front()
                {
                    assert(!empty, "Attempting to access the front of an empty byUTF");
                    return cast(dchar) buff;
                }

                void popFront() scope
                {
                    assert(!empty, "Attempting to popFront an empty byUTF");
                    if (r.empty)
                    {
                        buff = uint.max;
                    }
                    else
                    {
                        if (r.front < 0x80)
                        {
                            buff = r.front;
                            r.popFront;
                        }
                        else
                        {
                            buff = () @trusted { return decodeFront!(Yes.useReplacementDchar)(r); }();
                        }
                    }
                }

                static if (isForwardRange!R)
                {
                    @property auto save() return scope
                    {
                        auto ret = this;
                        ret.r = r.save;
                        return ret;
                    }
                }

                uint buff;
                R r;
            }

            return Result(r);
        }
        else
        {
            static struct Result
            {
                @property bool empty()
                {
                    return pos == fill && r.empty;
                }

                @property auto front() scope // 'scope' required by call to decodeFront() below
                {
                    if (pos == fill)
                    {
                        pos = 0;
                        auto c = r.front;

                        if (c <= 0x7F)
                        {
                            fill = 1;
                            r.popFront;
                            buf[pos] = cast(C) c;
                        }
                        else
                        {
                            static if (is(RC == dchar))
                            {
                                r.popFront;
                                dchar dc = c;
                            }
                            else
                                dchar dc = () @trusted { return decodeFront!(Yes.useReplacementDchar)(r); }();
                            fill = cast(ushort) encode!(Yes.useReplacementDchar)(buf, dc);
                        }
                    }
                    return buf[pos];
                }

                void popFront()
                {
                    if (pos == fill)
                        front;
                    ++pos;
                }

                static if (isForwardRange!R)
                {
                    @property auto save() return scope
                    /* `return scope` cannot be inferred because compiler does not
                     * track it backwards from assignment to local `ret`
                     */
                    {
                        auto ret = this;
                        ret.r = r.save;
                        return ret;
                    }
                }

            private:

                R r;
                C[4 / C.sizeof] buf = void;
                ushort pos, fill;
            }

            return Result(r);
        }
    }
}

auto test1()
{
    auto r = byUTF1!(dchar)(a);
    foreach (e; r)
    {
        res += cast(size_t) e;
    }
}

auto test2()
{
    auto r = byUTF2!(dchar)(a);
    foreach (e; r)
    {
        res += cast(size_t) e;
    }
}

void main()
{

    auto result  = to!Duration(benchmark!(test1)(testCount)[0]);
    writeln(res); // force execution
    auto result2  = to!Duration(benchmark!(test2)(testCount)[0]);
    writeln(res); // force execution

    writeln("old", "\t\t", result);
    writeln("new", "\t\t", result2);
}
```

</details>