@WalterBright 2019-03-13
Unfortunately, this has the constructor initiate reading the data, which can be unexpected.